backend/requirements.txt
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
markdown==3.4.4
django-filter==23.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
pillow==10.0.0
drf-writable-nested==0.7.0
python-dateutil==2.8.2